Fix Windows build error: align io_expander read_*_pins to uint16_t
The io_expander_ops_t interface declared read_input_pins/read_output_pins
with uint8_t * but the TCA9535 (16-bit I/O expander used on SenseCap
Indicator) implements them with uint16_t *. This caused an incompatible
function pointer error at sensecap_indicator_board.c:51.

Update the interface and the 8-bit expander implementations (tca9554,
ht8574) to use uint16_t * consistently. The 8-bit drivers now read into
a local uint8_t and widen it, preserving correct behaviour. This also
fixes the matching callers in sx126x_sensecap_board.c which already
declared their pin_val as uint16_t.

+++ components/i2c_devices/io_expander/ht8574.h
@@ -25,8 +25,8 @@ extern "C" {
 esp_err_t ht8574_init(uint8_t i2c_addr);
 esp_err_t ht8574_set_direction(uint8_t pin, bool is_output);
 esp_err_t ht8574_set_level(uint8_t pin, bool level);
-esp_err_t ht8574_read_output_pins(uint8_t *pin_val);
-esp_err_t ht8574_read_input_pins(uint8_t *pin_val);
+esp_err_t ht8574_read_output_pins(uint16_t *pin_val);
+esp_err_t ht8574_read_input_pins(uint16_t *pin_val);
 
 esp_err_t ht8574_multi_write_start(void);
 esp_err_t ht8574_multi_write_new_level(int pin, bool new_level);
